requirements.txt
Pillow>=10.0.0
numpy>=1.24.0
midiutil>=1.2.1
gradio>=4.0.0
matplotlib>=3.7.0
